requirements.txt
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@v0.7.4
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@v0.6
chumpy @ git+https://github.com/mattloper/chumpy
configer @ git+https://github.com/nghorbani/configer 
numpy
matplotlib
scipy
opencv-python
tqdm
PyYAML
tensorboard
torchgeometry
smplx
pyrender
open3d
hydra-core
imageio-ffmpeg
scikit-image
scikit-learn
joblib
cython
gdown
dill
motmetrics
scenedetect[opencv]
einops
mmcv
timm
xtcocotools
pandas
rerun-sdk>=0.9.0
